fix: add check for app type assertion, refactor: remove deploy package and move its components to app
