+++ grid-compose/cmd/down.go
@@ -10,7 +10,11 @@ var downCmd = &cobra.Command{
 	Use:   "down",
 	Short: "cancel your project on the grid",
 	Run: func(cmd *cobra.Command, args []string) {
-		app := cmd.Context().Value("app").(*app.App)
+		app, ok := cmd.Context().Value("app").(*app.App)
+		if !ok {
+			log.Fatal().Msg("app not found in context")
+		}
+
 		if err := app.Down(); err != nil {
 			log.Fatal().Err(err).Send()
 		}
