+++ grid-compose/cmd/up.go
@@ -10,7 +10,11 @@ var upCmd = &cobra.Command{
 	Use:   "up",
 	Short: "deploy application on the grid",
 	Run: func(cmd *cobra.Command, args []string) {
-		app := cmd.Context().Value("app").(*app.App)
+		app, ok := cmd.Context().Value("app").(*app.App)
+		if !ok {
+			log.Fatal().Msg("app not found in context")
+		}
+
 		if err := app.Up(cmd.Context()); err != nil {
 			log.Fatal().Err(err).Send()
 		}
