+++ grid-compose/internal/app/ps.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-client/workloads"
-	"github.com/threefoldtech/tfgrid-sdk-go/grid-compose/internal/deploy"
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-compose/internal/types"
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-compose/pkg/log"
 	"github.com/threefoldtech/zos/pkg/gridtypes"
@@ -26,7 +25,7 @@ func (a *App) Ps(ctx context.Context, verbose bool) error {
 	}
 
 	if len(a.Config.Networks) == 0 {
-		a.Config.Networks[deploy.GenerateDefaultNetworkName(a.Config.Services)] = types.Network{}
+		a.Config.Networks[a.GenerateDefaultNetworkName()] = types.Network{}
 	}
 
 	for networkName := range a.Config.Networks {
